Fix rate limit crash and Dapper int64 deserialization error
WordQueueProcessor: always use Haiku for batch generation. Using Sonnet
in the processor hit the 8k output tokens/min rate limit at lower account
tiers (5 concurrent x 4096 max_tokens = instant 429). Sonnet stays as
the live model for real-time user lookups (one at a time, no issue).
Failed words are reset to pending automatically on restart.

WordRepository.GetQueueStatusAsync: cast COUNT(*) to ::int in SQL.
PostgreSQL COUNT returns bigint (Int64) but QueueStatus record declares
int (Int32), causing Dapper deserialization to throw on every admin
queue/status call with a 500.

+++ api/Data/WordRepository.cs
@@ -278,10 +278,10 @@ public class WordRepository : IWordRepository
         using var conn = Connection();
         var sql = """
             SELECT
-                COUNT(*) FILTER (WHERE status = 'pending') as pending,
-                COUNT(*) FILTER (WHERE status = 'processing') as processing,
-                COUNT(*) FILTER (WHERE status = 'done') as done,
-                COUNT(*) FILTER (WHERE status = 'failed') as failed
+                COUNT(*) FILTER (WHERE status = 'pending')::int    as pending,
+                COUNT(*) FILTER (WHERE status = 'processing')::int as processing,
+                COUNT(*) FILTER (WHERE status = 'done')::int       as done,
+                COUNT(*) FILTER (WHERE status = 'failed')::int     as failed
             FROM word_queue
             """;
         return await conn.QuerySingleAsync<QueueStatus>(sql);
